fifo: Change index types to size_t
size_t represents the largest possible array index.

+++ inc/mcu-common/fifo.h
@@ -50,9 +50,9 @@ extern "C" {
 struct fifo {
 	volatile void *buffer;
 	size_t element_size;
-	int capacity;
-	volatile int tail;
-	volatile int head;
+	size_t capacity;
+	volatile size_t tail;
+	volatile size_t head;
 };
 
 bool fifo_init(struct fifo *fifo);
